Use dedicated hanlders class instead of traits for forms conditions
