+++ src/Glpi/Form/QuestionType/QuestionTypeEmail.php
@@ -35,15 +35,14 @@
 
 namespace Glpi\Form\QuestionType;
 
-use Glpi\Form\Condition\StringConditionTrait;
+use Glpi\Form\Condition\ConditionHandler\ConditionHandlerInterface;
+use Glpi\Form\Condition\ConditionHandler\StringConditionHandler;
 use Glpi\Form\Condition\UsedAsCriteriaInterface;
 use Override;
 use Session;
 
 final class QuestionTypeEmail extends AbstractQuestionTypeShortAnswer implements UsedAsCriteriaInterface
 {
-    use StringConditionTrait;
-
     #[Override]
     public function getInputType(): string
     {
@@ -66,5 +65,11 @@ final class QuestionTypeEmail extends AbstractQuestionTypeShortAnswer implements
     public function getWeight(): int
     {
         return 20;
+    }
+
+    #[Override]
+    public function getConditionHandler(): ConditionHandlerInterface
+    {
+        return new StringConditionHandler();
     }
 }
